Broke down Login Page to 3 components. Created a Logout Page. Updated HomePage with login and signup components. Added route to logout on navbar and app.js
