+++ client/src/components/SignUp/index.js
@@ -3,7 +3,7 @@ import 'whatwg-fetch';
 
 import {
   getFromStorage,
-  setInStorage,
+//   setInStorage,
 } from '../../utils/storage';
 
 class Home extends Component {
@@ -14,24 +14,26 @@ class Home extends Component {
       isLoading: true,
       token: '',
       signUpError: '',
-      signInError: '',
-      signInEmail: '',
-      signInPassword: '',
+    //   signInError: '',
+    //   signInEmail: '',
+    //   signInPassword: '',
+      signUpEmail: '',
+      signUpPassword: '',
       signUpFirstName: '',
       SignUpLastName: '',
     };
 
-    this.onTextboxChangeSignInEmail = this.onTextboxChangeSignInEmail.bind(this);
-    this.onTextboxChangeSignInPassword = this.onTextboxChangeSignInPassword.bind(this);
+    // this.onTextboxChangeSignInEmail = this.onTextboxChangeSignInEmail.bind(this);
+    // this.onTextboxChangeSignInPassword = this.onTextboxChangeSignInPassword.bind(this);
     this.onTextboxChangeSignUpEmail = this.onTextboxChangeSignUpEmail.bind(this);
     this.onTextboxChangeSignUpPassword = this.onTextboxChangeSignUpPassword.bind(this);
 
     this.onTextboxChangeSignUpFirstName = this.onTextboxChangeSignUpFirstName.bind(this);
     this.onTextboxChangeSignUpLastName = this.onTextboxChangeSignUpLastName.bind(this);
     
-    this.onSignIn = this.onSignIn.bind(this);
+    // this.onSignIn = this.onSignIn.bind(this);
     this.onSignUp = this.onSignUp.bind(this);
-    this.logout = this.logout.bind(this);
+    // this.logout = this.logout.bind(this);
   }
 
   componentDidMount() {
@@ -60,17 +62,17 @@ class Home extends Component {
     }
   }
 
-  onTextboxChangeSignInEmail(event) {
-    this.setState({
-      signInEmail: event.target.value,
-    });
-  }
+//   onTextboxChangeSignInEmail(event) {
+//     this.setState({
+//       signInEmail: event.target.value,
+//     });
+//   }
 
-  onTextboxChangeSignInPassword(event) {
-    this.setState({
-      signInPassword: event.target.value,
-    });
-  }
+//   onTextboxChangeSignInPassword(event) {
+//     this.setState({
+//       signInPassword: event.target.value,
+//     });
+//   }
 
   onTextboxChangeSignUpEmail(event) {
     this.setState({
@@ -144,86 +146,86 @@ class Home extends Component {
       });
   }
 
-  onSignIn() {
-    // Grab state
-    const {
-      signInEmail,
-      signInPassword,
-    } = this.state;
+//   onSignIn() {
+//     // Grab state
+//     const {
+//       signInEmail,
+//       signInPassword,
+//     } = this.state;
     
-    this.setState({
-      isLoading: true,
-    });
+//     this.setState({
+//       isLoading: true,
+//     });
 
-    // Post request to backend
-    fetch('/api/account/signin', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        email: signInEmail,
-        password: signInPassword,
-      }),
-    }).then(res => res.json())
-      .then(json => {
-        console.log('json', json);
-        if (json.success) {
-          setInStorage('the_main_app', { token: json.token });
-          //Redirect to auction list on success
-          window.location.pathname = '/auction';
-          // this.setState({
-          //   signInError: json.message,
-          //   isLoading: false,
-          //   signInPassword: '',
-          //   signInEmail: '',
-          //   token: json.token,
-          // });
-        } else {
-          this.setState({
-            signInError: json.message,
-            isLoading: false,
-          });
-        }
-      });
-  }
+//     // Post request to backend
+//     fetch('/api/account/signin', {
+//       method: 'POST',
+//       headers: {
+//         'Content-Type': 'application/json'
+//       },
+//       body: JSON.stringify({
+//         email: signInEmail,
+//         password: signInPassword,
+//       }),
+//     }).then(res => res.json())
+//       .then(json => {
+//         console.log('json', json);
+//         if (json.success) {
+//           setInStorage('the_main_app', { token: json.token });
+//           //Redirect to auction list on success
+//           window.location.pathname = '/auction';
+//           // this.setState({
+//           //   signInError: json.message,
+//           //   isLoading: false,
+//           //   signInPassword: '',
+//           //   signInEmail: '',
+//           //   token: json.token,
+//           // });
+//         } else {
+//           this.setState({
+//             signInError: json.message,
+//             isLoading: false,
+//           });
+//         }
+//       });
+//   }
 
-  logout() {
-    this.setState({
-      isLoading: true,
-    });
-    const obj = getFromStorage('the_main_app');
-    if (obj && obj.token) {
-      const { token } = obj;
-      // Verify token
-      fetch('/api/account/logout?token=' + token)
-        .then(res => res.json())
-        .then(json => {
-          if (json.success) {
-            this.setState({
-              token: '',
-              isLoading: false
-            });
-          } else {
-            this.setState({
-              isLoading: false,
-            });
-          }
-        });
-    } else {
-      this.setState({
-        isLoading: false,
-      });
-    }
-  }
+//   logout() {
+//     this.setState({
+//       isLoading: true,
+//     });
+//     const obj = getFromStorage('the_main_app');
+//     if (obj && obj.token) {
+//       const { token } = obj;
+//       // Verify token
+//       fetch('/api/account/logout?token=' + token)
+//         .then(res => res.json())
+//         .then(json => {
+//           if (json.success) {
+//             this.setState({
+//               token: '',
+//               isLoading: false
+//             });
+//           } else {
+//             this.setState({
+//               isLoading: false,
+//             });
+//           }
+//         });
+//     } else {
+//       this.setState({
+//         isLoading: false,
+//       });
+//     }
+//   }
 
   render() {
     const {
       isLoading,
       token,
-      signInError,
-      signInEmail,
-      signInPassword,
+    //   signInError,
+    //   signInEmail,
+    //   signInPassword,
       signUpFirstName,
       SignUpLastName,
       signUpEmail,
@@ -239,12 +241,12 @@ class Home extends Component {
       return (
         <div>
           <div>
-            {
+            {/* {
               (signInError) ? (
                 <p>{signInError}</p>
               ) : (null)
-            }
-            <p>Sign In</p>
+            } */}
+            {/* <p>Sign In</p>
             <input
               type="email"
               placeholder="Email"
@@ -263,7 +265,7 @@ class Home extends Component {
           </div>
           <br />
           <br />
-          <div>
+          <div> */}
             {
               (signUpError) ? (
                 <p>{signUpError}</p>
@@ -301,12 +303,12 @@ class Home extends Component {
       );
     }
 
-    return (
-      <div>
-        <p>Press the button below to confirm your logout</p>
-        <button onClick={this.logout}>Logout</button>
-      </div>
-    );
+    // return (
+    //   <div>
+    //     <p>Press the button below to confirm your logout</p>
+    //     <button onClick={this.logout}>Logout</button>
+    //   </div>
+    // );
   }
 }
