+++ client/src/pages/HomePage.js
@@ -1,18 +1,20 @@
 import React, { Component } from "react";
-import Nav from "../components/Nav";
-import Footer from "../components/Footer";
-import Home from "../components/Home";
+// import Home from "../components/Home";
+import SignIn from "../components/SignIn";
+import SignUp from "../components/SignUp";
 
 class HomePage extends Component {
 //Might have to add parts from Home.js to the HomePage file...
     render(){
         return(
             <>
-            <Home />
+            <SignIn />
+            <br />
+            <br />
+            <SignUp />
             </>
-        )
+        );
     }
-
 }
 
 export default HomePage;
