+++ client/src/pages/LogOutPage.js
@@ -1,0 +1,14 @@
+import React, { Component } from "react";
+import LogOut from "../components/LogOut";
+
+class LogOutPage extends Component {
+    render(){
+        return(
+            <>
+            <LogOut />
+            </>
+        );
+    }
+}
+
+export default LogOutPage;
